+++ src/main/kotlin/com/zp4rker/discore/console/ConsoleCommand.kt
@@ -1,0 +1,12 @@
+package com.zp4rker.discore.console
+
+/**
+ * @author zp4rker
+ *
+ * Console command interface.
+ */
+interface ConsoleCommand {
+
+    fun handleCommand(command: String)
+
+}